Fix compiler warning
git-svn-id: http://root.cern.ch/svn/root/trunk@43533 27541ba8-7e3a-0410-8455-c3a389f83636

+++ core/utils/src/rootcling.cxx
@@ -5078,7 +5078,7 @@ static ESourceFileKind GetSourceFileKind(const char* filename)
    }
    } // switch extension length
 
-   static const int lenLinkdefdot = 8;
+   static const size_t lenLinkdefdot = 8;
    if (ret == kSFKHeader && len - lenExt >= lenLinkdefdot) {
       if ((strstr(filename,"LinkDef") || strstr(filename,"Linkdef") ||
            strstr(filename,"linkdef")) && strstr(filename,".h")) {
